feat: separated quote components
removed reference warning component since it is no longer needed
